requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
reportlab==3.6.1
slackclient==2.9.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7 # For swagger
pillow~=10.3.0
pip~=24.0
wheel~=0.43.0
pytz~=2024.1
PyYAML~=6.0.1
idna~=3.7
requests~=2.32.3
packaging~=24.0
inflection~=0.5.1
uritemplate~=4.1.1
setuptools~=69.5.1
django-filter==24.2